+++ jslint_ci.sh
@@ -1015,8 +1015,7 @@ shGithubCheckoutRemote() {(set -e
     rm -rf __tmp1
     git reset "origin/$GITHUB_REF_NAME" --hard
     # fetch jslint_ci.sh from trusted source
-    git branch -D __tmp1 &>/dev/null || true
-    shGitCmdWithGithubToken fetch origin alpha:__tmp1 --depth=1
+    shGitCmdWithGithubToken fetch origin alpha:__tmp1 --depth=1 -f
     for FILE in .ci.sh .ci2.sh jslint_ci.sh myci2.sh
     do
         if [ -f "$FILE" ]