+++ sqlmath.mjs
@@ -60,6 +60,7 @@ let SQLITE_OPEN_TEMP_JOURNAL = 0x00001000;  /* VFS only */
 let SQLITE_OPEN_TRANSIENT_DB = 0x00000400;  /* VFS only */
 let SQLITE_OPEN_URI = 0x00000040;           /* Ok for sqlite3_open_v2() */
 let SQLITE_OPEN_WAL = 0x00080000;           /* VFS only */
+let SQLITE_WIN_SLR_ELEM_SIZE = 10;
 let cModule;
 let cModulePath;
 let consoleError = console.error;
@@ -97,7 +98,7 @@ let {
 let sqlMessageDict = {}; // dict of web-worker-callbacks
 let sqlMessageId = 0;
 let sqlWorker;
-let version = "v2023.7.1-beta";
+let version = "v2023.7.21";
 
 function assertJsonEqual(aa, bb, message) {
 
@@ -1356,6 +1357,7 @@ export {
     SQLITE_OPEN_TRANSIENT_DB,
     SQLITE_OPEN_URI,
     SQLITE_OPEN_WAL,
+    SQLITE_WIN_SLR_ELEM_SIZE,
     assertJsonEqual,
     assertNumericalEqual,
     assertOrThrow,